fix: resolve compile and layout scope errors on ChatScreen and ChatMessageList

+++ android-compose/feature-chat/src/main/java/com/letta/mobile/feature/chat/screen/ChatMessageList.kt
@@ -1057,7 +1057,7 @@ internal fun ChatMessageList(
                                     val offset = nextPromptItemInfo.offset
                                     val size = nextPromptItemInfo.size
                                     val yTopNext = viewportHeight - (offset + size)
-                                    val pushOffset = yTopNext - stickyHeaderHeight
+                                    val pushOffset = (yTopNext - stickyHeaderHeight).toFloat()
                                     if (pushOffset < yOffset) {
                                         yOffset = pushOffset
                                     }
